+++ internal/pkg/plan/dag.go
@@ -473,7 +473,7 @@ func (g *Dag) walk(down bool, processCh chan<- NamedNode, doneCh chan NamedNode)
 				}
 
 				if len(inProgressNodes) > 0 {
-					_, _ = printer.Fprintf("[yellow]Waiting on: %s...\n", strings.Join(inProgressNodes, ", "))
+					_, _ = printer.Fprintf("[yellow]Waiting on: [bold]%s[reset][yellow]...\n", strings.Join(inProgressNodes, ", "))
 				}
 			}
 		}