+++ internal/pkg/stack/loader.go
@@ -96,8 +96,8 @@ func BuildStack(stackName string, stackFilePath string, cliStackConfig *structs.
 			"define it in your stack.")
 	}
 
-	_, err = printer.Fprintf("Stack config file '[bold]%s[reset]' contains %d "+
-		"manifest(s) and %d kapp(s) in total.\n\n", stackFilePath, len(stackConfig.Manifests()), numKapps)
+	_, err = printer.Fprintf("Stack '[bold]%s[reset]' in config file '[bold]%s[reset]' contains %d "+
+		"manifest(s) and %d kapp(s) in total.\n\n", stackName, stackFilePath, len(stackConfig.Manifests()), numKapps)
 	if err != nil {
 		return nil, errors.WithStack(err)
 	}